dsl_compiler/requirements.txt
# DSL Compiler Dependencies

# Core dependencies
pydantic>=2.0.0              # Data validation and serialization
ruamel.yaml>=0.17.0          # YAML processing with comments preservation
python-dotenv>=1.0.0         # Environment variable loading

# CLI dependencies
typer>=0.9.0                 # Command line interface framework
rich>=13.0.0                 # Rich text and beautiful formatting for terminal

# HTTP client for LLM integration
aiohttp>=3.8.0               # Async HTTP client

# Optional dependencies for development
pytest>=7.0.0                # Testing framework
pytest-asyncio>=0.21.0       # Async testing support
black>=23.0.0                # Code formatting
flake8>=6.0.0                # Code linting
mypy>=1.0.0                  # Type checking

# Documentation
sphinx>=6.0.0                # Documentation generation
sphinx-rtd-theme>=1.3.0      # Read the Docs theme 